requirements.txt
ctransformers
sentence-transformers
pinecone
langchain
langchain_pinecone
flask
pypdf
python-dotenv
-e .
